Rework how the gpodder client receives and passes along credentials to other request classes to avoid having to make the consumer keep track of the user name

+++ GPodder.NET/Models/Device.cs
@@ -35,6 +35,6 @@ namespace GPodder.NET.Models
         /// Gets or sets the number of subscriptions on the device.
         /// </summary>
         [JsonPropertyName("subscriptions")]
-        public int Subscrciptions { get; set; }
+        public long Subscrciptions { get; set; }
     }
 }
